requirements.txt
crewai>=0.28.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
jupyter>=1.0.0
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
restrictedpython>=6.0
black>=23.0.0
flake8>=6.0.0
python-docx>=1.0.0
markdown2>=2.4.0
Pillow>=10.0.0